+++ CHANGELOG.md
@@ -57,6 +57,7 @@ All notable changes to this project will be documented here. The format is based
 - The simulator panel will now only render the UI once one simulation state has been emitted from the simulator (#589)
 - Physical offset frames can now be translated+rotated in the model editor tab's 3D viewport (#583)
 - Wrap objects can now be translated+rotated in the model editor tab's 3D viewport (#588)
+- The "Reload [Model]" button now forces OSC to also reload any mesh files (#594)
 
 ## [0.3.2] - 2023/01/09